Refactor🔨: String모아둔 Model생성
