--- Jikyeojulge/Model/MainViewComponent.swift
@@ -1,15 +1,0 @@
-//
-//  MainViewComponent.swift
-//  Jikyeojulge
-//
-//  Created by 지준용 on 2022/07/09.
-//
-
-import Foundation
-
-struct MainViewComponent {
-    let name =  "이름"
-    let birth = "생년월일"
-    let bloodType = "혈액형"
-    let contact = "비상연락처"
-}
