+++ Jikyeojulge/Model/ViewStirng.swift
@@ -1,11 +1,18 @@
 //
-//  OnboardingComponent.swift
+//  MainViewComponent.swift
 //  Jikyeojulge
 //
-//  Created by 지준용 on 2022/07/07.
+//  Created by 지준용 on 2022/07/09.
 //
 
 import Foundation
+
+struct MainViewComponent {
+    let name =  "이름"
+    let birth = "생년월일"
+    let bloodType = "혈액형"
+    let contact = "비상연락처"
+}
 
 struct OnboardingComponent {
     let title = ["1단계", "2단계", "3단계", "4단계", "5단계", "6단계"]
